+++ packages/eslint-plugin-query/src/rules/exhaustive-deps/exhaustive-deps.rule.ts
@@ -82,6 +82,7 @@ export const rule = createRule({
             reference,
             scopeManager,
             node: getQueryFnRelevantNode(queryFn),
+            filename: context.filename,
           }),
         )