+++ packages/query-devtools/src/Devtools.tsx
@@ -76,7 +76,6 @@ import type {
   Query,
   QueryCache,
   QueryCacheNotifyEvent,
-  QueryState,
 } from '@tanstack/query-core'
 import type { StorageObject, StorageSetter } from '@solid-primitives/storage'
 import type { Accessor, Component, JSX, Setter } from 'solid-js'
@@ -1957,13 +1956,14 @@ const QueryDetails = () => {
     const __previousQueryOptions = activeQueryVal.options
 
     activeQueryVal.setState({
+      data: undefined,
       status: 'error',
       error,
       fetchMeta: {
         ...activeQueryVal.state.fetchMeta,
         __previousQueryOptions,
       } as any,
-    } as QueryState<unknown, Error>)
+    })
   }
 
   const restoreQueryAfterLoadingOrError = () => {
@@ -2199,7 +2199,7 @@ const QueryDetails = () => {
                     ...activeQueryVal.state.fetchMeta,
                     __previousQueryOptions,
                   } as any,
-                } as QueryState<unknown, Error>)
+                })
               }
             }}
           >